doctor_appointment/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
celery
redis
psycopg2-binary
paypalrestsdk
python-dotenv
gunicorn
